Add context menu module
